refactor(paseo): satisfy review module limits and bound health probes
