+++ ci/test-suite-ratchet-baseline.json
@@ -1,5 +1,5 @@
 {
   "schema_version": 1,
-  "maximum_collected_cases": 16312,
-  "maximum_test_source_lines": 351613
+  "maximum_collected_cases": 16316,
+  "maximum_test_source_lines": 351667
 }
