+++ src/codex_plugin_scanner/guard/daemon/client.py
@@ -3,12 +3,15 @@
 from __future__ import annotations
 
 import json
+import socket
+import time
 import urllib.error
 import urllib.request
 from collections.abc import Mapping
-from contextlib import closing
+from contextlib import closing, suppress
 from http.client import HTTPConnection, HTTPException
 from pathlib import Path
+from threading import Timer
 from typing import TypeGuard
 from urllib.parse import urlsplit
 
@@ -19,6 +22,15 @@ from .manager import (
     load_guard_daemon_url,
     load_running_guard_daemon_identity,
 )
+
+_HEALTH_PROBE_DEADLINE_SECONDS = 1.0
+
+
+def _interrupt_health_socket(stream: socket.socket) -> None:
+    """Interrupt a blocked header/body read when the whole probe deadline expires."""
+    # A completed request may already have closed its socket.
+    with suppress(OSError):
+        stream.shutdown(socket.SHUT_RDWR)
 
 
 def read_guard_health_details(daemon_url: str, auth_token: str) -> dict[str, object] | None:
@@ -39,12 +51,28 @@ def read_guard_health_details(daemon_url: str, auth_token: str) -> dict[str, obj
             return None
         # HTTPConnection neither consults proxy environment variables nor follows
         # redirects. Never forward the daemon token to a redirected authority.
-        with closing(HTTPConnection(parsed.hostname, parsed.port, timeout=1.0)) as connection:
-            connection.request("GET", "/v1/healthz/details", headers={"X-Guard-Token": auth_token})
-            response = connection.getresponse()
-            if response.status != 200:
+        deadline = time.monotonic() + _HEALTH_PROBE_DEADLINE_SECONDS
+        with closing(
+            HTTPConnection(parsed.hostname, parsed.port, timeout=_HEALTH_PROBE_DEADLINE_SECONDS)
+        ) as connection:
+            connection.connect()
+            stream = connection.sock
+            remaining = deadline - time.monotonic()
+            if stream is None or remaining <= 0:
                 return None
-            content = response.read(65_537)
+            interrupt = Timer(remaining, _interrupt_health_socket, args=(stream,))
+            interrupt.daemon = True
+            interrupt.start()
+            try:
+                connection.request("GET", "/v1/healthz/details", headers={"X-Guard-Token": auth_token})
+                response = connection.getresponse()
+                if response.status != 200:
+                    return None
+                content = response.read(65_537)
+            finally:
+                interrupt.cancel()
+        if time.monotonic() >= deadline:
+            return None
         if len(content) > 65_536:
             return None
         payload = json.loads(content.decode("utf-8"))
